@@ -10,37 +10,10 @@ use Yiisoft\Profiler\Message;
 use function dirname;
 
 /**
- * FileTarget records profiling messages in a file specified via {@see FileTarget::$filename}.
+ * FileTarget records profiling messages in a file specified via {@see FileTarget::$filePath}.
  */
 final class FileTarget extends AbstractTarget
 {
-    /**
-     * @var string Path of the file to write to. It may contain the placeholders,
-     * which will be replaced by computed values. The supported placeholders are:
-     *
-     * - '{ts}' - profiling completion timestamp.
-     * - '{date}' - profiling completion date in format 'ymd'.
-     * - '{time}' - profiling completion time in format 'His'.
-     *
-     * The directory containing the file will be automatically created if not existing.
-     * If target file is already exist it will be overridden.
-     */
-    private string $filename;
-
-    /**
-     * @var int The permission to be set for newly created directories.
-     * This value will be used by PHP chmod() function. No umask will be applied.
-     * Defaults to 0775, meaning the directory is read-writable by owner and group,
-     * but read-only for other users.
-     */
-    private int $directoryMode;
-
-    /**
-     * @var float Time of the beginning of the request. Can be set as `microtime(true)` or
-     * `$_SERVER['REQUEST_TIME_FLOAT']` in config.
-     */
-    private float $requestBeginTime;
-
     /**
      * @param string $filePath Path of the file to write to. It may contain the placeholders,
      * which will be replaced by computed values. The supported placeholders are:
@@ -58,11 +31,11 @@ final class FileTarget extends AbstractTarget
      * Defaults to 0775, meaning the directory is read-writable by owner and group,
      * but read-only for other users.
      */
-    public function __construct(string $filePath, float $requestBeginTime, int $directoryMode = 0775)
-    {
-        $this->filename = $filePath;
-        $this->requestBeginTime = $requestBeginTime;
-        $this->directoryMode = $directoryMode;
+    public function __construct(
+        private string $filePath,
+        private float $requestBeginTime,
+        private int $directoryMode = 0775
+    ) {
     }
 
     public function export(array $messages): void
@@ -99,17 +72,14 @@ final class FileTarget extends AbstractTarget
         return preg_replace_callback(
             '/{\\w+}/',
             static function (array $matches) {
-                switch ($matches[0]) {
-                    case '{ts}':
-                        return (string) time();
-                    case '{date}':
-                        return gmdate('ymd');
-                    case '{time}':
-                        return gmdate('His');
-                }
-                return $matches[0];
+                return match ($matches[0]) {
+                    '{ts}' => (string)time(),
+                    '{date}' => gmdate('ymd'),
+                    '{time}' => gmdate('His'),
+                    default => $matches[0],
+                };
             },
-            $this->filename
+            $this->filePath
         );
     }
 
