@@ -14,44 +14,22 @@ use Psr\Log\LogLevel;
 final class Message
 {
     /**
-     * @var string Message level.
-     *
-     * @see LogLevel See constants for valid level names.
-     */
-    private string $level;
-
-    /**
-     * @var string Message token.
-     */
-    private string $token;
-
-    /**
-     * @var array Message context.
-     *
-     * Message context has a following keys:
+     * @param string $level Message level.
+     * @param string $token Message token.
+     * @param array $context Message context with a following keys:
      *
      * - category: string, message category.
      * - memory: int, memory usage in bytes, obtained by `memory_get_usage()`.
      * - time: float, message timestamp obtained by microtime(true).
      * - trace: array, debug backtrace, contains the application code call stacks.
-     */
-    private array $context;
-
-    /**
-     * @param string $level Message level.
-     * @param string $token Message token.
-     * @param array $context Message context.
      *
      * @throws InvalidArgumentException for invalid log message level.
      *
      * @see LoggerTrait::log()
      * @see LogLevel
      */
-    public function __construct(string $level, string $token, array $context = [])
+    public function __construct(private string $level, private string $token, private array $context = [])
     {
-        $this->level = $level;
-        $this->token = $token;
-        $this->context = $context;
     }
 
     /**
@@ -82,17 +60,15 @@ final class Message
      * @param string|null $name The context parameter name.
      * @param mixed $default If the context parameter does not exist, the `$default` will be returned.
      *
-     * @return mixed The context parameter value.
-     *
      * @psalm-return ($name is null ? array : mixed|null)
      */
-    public function context(string $name = null, $default = null)
+    public function context(string $name = null, mixed $default = null): mixed
     {
         if ($name === null) {
             return $this->context;
         }
 
-        /** @psalm-suppress MixedReturnStatement */
+        /** @var array<array-key, mixed> */
         return $this->context[$name] ?? $default;
     }
 }
